push

+++ src/mtm/db/Interface/UserInterface.java
@@ -317,52 +317,60 @@ public class UserInterface
 			jdbcManager.insert(hosp);
 			jdbcManager.setHospitalID(hosp);
 			
-			System.out.println("Do you want torelate an order with this hospital. YES or NO:\n");
+			System.out.println("Do you want to relate an order with this hospital. YES or NO:\n");
 			String option = writeString();
 			if(writeOption(option))
 			{
-				boolean keepRelating = true;
-				while(keepRelating)
+				if(jdbcManager.valExist("SELECT * FROM instrument WHERE instrument_ID = ?",1,null))
 				{
-					System.out.println("The Order allready exixt?. YES or NO");
-					option = writeString();
-					if(writeOption(option))
+					boolean keepRelating = true;
+					while(keepRelating)
 					{
-						System.out.println("Select one of the Orders");
-						listOrders(false);
-						int op2 = writeNumber();
-						System.out.println("Insert the amountOrder");
-						int tao = writeNumber();
-						jdbcManager.setRelationHospitalOrder(hosp.getHospitalID(),op2,tao);
-					}else
-					{
-				    	System.out.println("Introduce the values:\n");
-				    	
-						Order ord = createOrder();
-						jdbcManager.insert(ord);
-						jdbcManager.setOrderID(ord);
-						System.out.println("Select the Primary Key of the instrument you want to order\n");
-				    	listInstruments(false);
-				    	int opt = writeNumber();
-				    	
-						System.out.println("Insert the amountOrder");
-						int tao = writeNumber();
+						System.out.println("The Order allready exixt?. YES or NO");
+						option = writeString();
+						if(writeOption(option))
+						{
+							System.out.println("Select one of the Orders");
+							listOrders(false);
+							int op2 = writeNumber();
+							System.out.println("Insert the amountOrder");
+							int tao = writeNumber();
+							jdbcManager.setRelationHospitalOrder(hosp.getHospitalID(),op2,tao);
+						}else
+						{
+					    	System.out.println("Introduce the values:\n");
+					    	
+							Order ord = createOrder();
+
+							jdbcManager.setOrderID(ord);
+							System.out.println("Select the Primary Key of the instrument you want to order\n");
+					    	listInstruments(false);
+					    	int opt = writeNumber();
+					    	
+							System.out.println("Insert the amountOrder");
+							int tao = writeNumber();
+							
+							jdbcManager.setRelationHospitalOrder(hosp.getHospitalID(),ord.getOrderID(),tao);
+							jdbcManager.setRelationInstrumentOrder(opt,ord.getOrderID());
+						}
+						System.out.println("Do you want to keep relating? YES,NO\n");
+						option = writeString();
+						if(option.equals("NO"))
+						{
+							keepRelating = false;
+						}
 						
-						jdbcManager.setRelationHospitalOrder(hosp.getHospitalID(),ord.getOrderID(),tao);
-						jdbcManager.setRelationInstrumentOrder(opt,ord.getOrderID());
 					}
-					System.out.println("Do you want to keep relating? YES,NO\n");
-					option = writeString();
-					if(option.equals("NO"))
-					{
-						keepRelating = false;
-					}
-					
+				}else
+				{
+					System.out.println("No intruments exist, therefore, you cant perform an order");
 				}
+				
 			}else
 			{
 				System.out.println("No Order will be related");
 			}
+			
 			break;
 		case 4: //Instrument
 			Instrument inst = createInstrument();
@@ -605,7 +613,7 @@ public class UserInterface
 		System.out.println("Size of machinery");
 		int d=writeNumber();
 
-		//mach = jdbcManager.createPojoMachinery(a,b,c1[0],c1[1],c1[2],d);
+		mach = jdbcManager.createPojoMachinery(a,b,c1[0],c1[1],c1[2],d);
 		
 		return mach;
 }
